% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaresp.r
\name{betaresp}
\alias{betaresp}
\alias{print.betaresp}
\alias{plot.betaresp}
\alias{betaresponse}
\title{Beta response models}
\usage{
betaresp(x, p1, p2, alpha, gamma, hi)
}
\arguments{
\item{x}{Gradient values}

\item{p1, p2}{response endpoints}

\item{alpha, gamma}{shape parameters of the response}

\item{hi}{maximum height of response}
}
\description{
Generating beta response models. Implemented to compare the hierarchical logistic regression models
  with beta response models.
}
\examples{
 x <- seq(0,3, len=201)
 m <- betaresp(x, 2, 0.8, 0.1, 0.5, 0.8)
 plot(m)

}
\author{
Jari Oksanen
}
