% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eHOF-package.R
\docType{package}
\name{eHOF-package}
\alias{eHOF-package}
\title{eHOF: Extended HOF (Huisman-Olff-Fresco) Models}
\description{
Extended and enhanced hierarchical logistic regression models (called Huisman-Olff-Fresco in biology, see Huisman et al. 1993 Journal of Vegetation Science \doi{10.1111/jvs.12050}) models. Response curves along one-dimensional gradients including no response, monotone, plateau, unimodal and bimodal models.
}
\keyword{internal}
