% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eHOF-package.R
\encoding{UTF-8}
\name{eHOF}
\alias{eHOF}
\title{Extended HOF (Huisman-Olff-Fresco) Models}
\description{
Extended and enhanced hierarchical logistic regression models (called Huisman-Olff-Fresco in biology, see Huisman et al. 1993 Journal of Vegetation Science <doi:10.1111/jvs.12050>) models. Response curves along one-dimensional gradients including no response, monotone, plateau, unimodal and bimodal models.
}
\references{
Jansen, F., Dengler, J (2011) Plant names in vegetation databases - a neglected source of bias,
Journal of vegetation science, 21(6), 1179-1186. http://dx.doi.org/10.1111/j.1654-1103.2010.01209.x
}
