\name{HOF-internal}
\alias{HOF.fun}
\alias{logLik.HOF}
\alias{Para_niche}
\alias{Para_opt}
\alias{scale01}
\alias{svHOF}
\alias{HOF.model}
\alias{[.HOF.list}
\alias{eHOF.modelnames}

\title{Internal eHOF functions}

\description{
  Internal eHOF functions that are not intended to be called directly,
  but only within other functions.
}

\keyword{internal}
