\encoding{UTF-8}
\name{HOF-package}
\alias{HOF-package}
\docType{package}

\title{Package for hierarchical logistic regression analysis.}

\description{
The HOF package provides functions for hierachical logistic regressions, extended to a bimodal model and a bootstrap approach to guarantee model stability. In ecology such hierarchical models are known as HOF models (Huisman-Olff-Fresco 1993).}

\details{
\tabular{ll}{Package: \tab HOF\cr
Title: \tab Extended Huisman-Olff-Fresco-Models\cr
Author: \tab Florian Jansen <jansen@uni-greifswald.de>\cr
\tab Jari Oksanen\cr
Maintainer: \tab Florian Jansen <jansen@uni-greifswald.de>\cr
Depends: \tab R (>= 2.0.0), \code{\link[vegdata]{vegdata}}\cr
Suggests: \tab \code{\link[vegan]{vegan}}\cr
URL: \tab http://geobot.botanik.uni-greifswald.de/download\cr}
}

\author{Florian Jansen <jansen@uni-greifswald.de> and 
    Jari Oksanen
  }

\seealso{\code{\link{HOF}}}

\references{
  Huisman, J., Olff, H. & Fresco, L.F.M. (1993). A hierarchical set of models for species response analysis. \emph{Journal of Vegetation Science} 4, 37-46.
}

\keyword{ package }


