% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autolayout.r
\name{autolayout}
\alias{autolayout}
\title{Automatically Specifying Complex Plot Arrangements}
\usage{
autolayout(N, byrow = TRUE, ...)
}
\arguments{
\item{N}{number of subfigures}

\item{byrow}{see \link{matrix}}

\item{...}{additional arguments for \link{layout} or \link{matrix}.}
}
\value{
Number of figures, N, see \link{layout}.
}
\description{
Automatic version of function \code{\link{layout}} for given number of subplots. Defined matrix schemes for \code{\link{layout}} for up to 32 subfigures.
}
\examples{
for(i in sample(1:32, 1)) {
 autolayout(i, byrow=TRUE)
 layout.show(i)
 }

}
