\name{HOF-internal}
\alias{coef.HOF}
\alias{deviance.HOF}
\alias{HOF.fun}
\alias{logLik.HOF}
\alias{Para.niche} 
\alias{Para.opt}
\alias{Para.deriv}
\alias{scale01} 
\alias{svHOF}
\alias{HOF.model}
\alias{[.HOF.list}
\alias{eHOF.modelnames}

\title{Internal eHOF functions}

\description{
  Internal eHOF functions that are not intended to be called directly,
  but only within other functions.
}

\keyword{internal}